+++ components/src/main/java/com/mercadolibre/android/andesui/dropdown/factory/AndesDropdownConfigurationFactory.kt
@@ -2,11 +2,11 @@ package com.mercadolibre.android.andesui.dropdown.factory
 
 import android.content.Context
 import com.mercadolibre.android.andesui.dropdown.state.AndesDropdownStateInterface
-import com.mercadolibre.android.andesui.dropdown.type.AndesDropdownMenuType
+import com.mercadolibre.android.andesui.dropdown.type.AndesDropdownMenuInterface
 import com.mercadolibre.android.andesui.textfield.state.AndesTextfieldState
 
 internal data class AndesDropdownConfiguration(
-    val menuType: AndesDropdownMenuType,
+    val menuType: AndesDropdownMenuInterface,
     val label: String?,
     val helper: String?,
     val placeHolder: String?,
@@ -19,7 +19,7 @@ internal object AndesDropdownConfigurationFactory {
     fun create(context: Context, andesDropdownAttrs: AndesDropdownAttrs): AndesDropdownConfiguration {
         return with(andesDropdownAttrs) {
             AndesDropdownConfiguration(
-                    menuType = andesDropdownAttrs.andesDropdownMenuType,
+                    menuType = andesDropdownAttrs.andesDropdownMenuType.type,
                     label = andesDropdownAttrs.andesDropdownLabel,
                     helper = andesDropdownAttrs.andesDropdownHelper,
                     placeHolder = andesDropdownAttrs.andesDropdownPlaceHolder,